wired call to the database in actions and reducers
relates to#32
