+++ src/client/components/pages/PostJob.js
@@ -3,47 +3,20 @@ import {connect} from 'react-redux';
 import Form_Post_Job from '../Form_Post_Job';
 import {sendingRequest} from '../../actions/AppActions';
 import LoadingIndicator from '../LoadingIndicator.js';
-import * as actions from '../../actions/post_job.js';
-import axios from 'axios';
 
-const testObject = {
-    start_date: '01/12/90',
-    start_time: '12:00',
-    end_date: '12/12/12',
-    end_time: '07:00',
-    job_title: 'Hello Chello',
-    description: 'Lorem Ipsum',
-    rate: '100',
-    resident_id: '1',
-    category: 'photography'
-};
+// const testObject = {
+//     start_date: '01/12/90',
+//     start_time: '12:00',
+//     end_date: '12/12/12',
+//     end_time: '07:00',
+//     job_title: 'Hello Chello',
+//     description: 'Lorem Ipsum',
+//     rate: '100',
+//     resident_id: '1',
+//     category: 'photography'
+// };
 
 class PostJob extends Component {
-    constructor() {
-        super();
-    }
-
-    componentWillMount() {
-        //this.props.postJobs
-        axios
-            .post('/api/jobs', {
-                start_date: '03/07/2017',
-                start_time: '12:00',
-                end_date: '12/12/12',
-                end_time: '07:00',
-                job_title: 'Hello Chello',
-                description: 'Lorem Ipsum',
-                rate: '100.00',
-                resident_id: '1',
-                category: 'horse riding'
-            })
-            .then(function(response) {
-                console.log(response);
-            })
-            .catch(function(error) {
-                console.log(error);
-            });
-    }
     render() {
         // const dispatch = this.props.dispatch;
         // const {currentlySending} = this.props.data;
@@ -62,8 +35,4 @@ class PostJob extends Component {
     }
 }
 
-function mapStateToProps(state) {
-    return {};
-}
-
-export default connect(mapStateToProps, actions)(PostJob);
+export default PostJob;
